+++ samples/Axent.ExampleApi/ExampleRequestPipe.cs
@@ -11,7 +11,7 @@ internal sealed class ExampleRequestPipe<TRequest, TResponse> : IAxentPipe<TRequ
         _logger = logger;
     }
 
-    public Task<Response<TResponse>> ProcessAsync(IPipelineChain<TRequest, TResponse> chain, RequestContext<TRequest> context, CancellationToken cancellationToken = default)
+    public ValueTask<Response<TResponse>> ProcessAsync(IPipelineChain<TRequest, TResponse> chain, RequestContext<TRequest> context, CancellationToken cancellationToken = default)
     {
         _logger.LogInformation("This pipe runs during every request.");
         return chain.NextAsync(context, cancellationToken);