+++ samples/Axent.ExampleApi/Program.cs
@@ -1,12 +1,14 @@
 using Axent.Abstractions;
 using Axent.Core;
+using Axent.Core.DependencyInjection;
 using Axent.ExampleApi;
 using Axent.Extensions.AspNetCore;
 
 var builder = WebApplication.CreateBuilder(args);
 
-builder.Services.AddAxent()
-    .AddRequestHandlers(AssemblyProvider.Current)
+builder.Services.AddAxent(o => builder.Configuration.Bind("AppSettings:Axent", o))
+    .AddTracing()
+    .AddRequestHandlersFromAssembly(AssemblyProvider.Current)
     .AddPipe<OtherRequestPipe>()
     .AddPipe(typeof(ExampleRequestPipe<,>));