+++ src/Axent.Abstractions/IPipelineChain.cs
@@ -5,7 +5,7 @@ public interface IPipelineChain<TRequest, TResponse>
     /// <summary>
     /// Moves to the next pipe in the chain.
     /// </summary>
-    Task<Response<TResponse>> NextAsync(
+    ValueTask<Response<TResponse>> NextAsync(
         RequestContext<TRequest> context,
         CancellationToken cancellationToken = default);
 }