+++ src/Axent.Abstractions/ISender.cs
@@ -2,5 +2,5 @@ namespace Axent.Abstractions;
 
 public interface ISender
 {
-    Task<Response<TResponse>> SendAsync<TResponse>(IRequest<TResponse> request, CancellationToken cancellationToken = default);
+    ValueTask<Response<TResponse>> SendAsync<TResponse>(IRequest<TResponse> request, CancellationToken cancellationToken = default);
 }