+++ src/Axent.Abstractions/Unit.cs
@@ -1,4 +1,4 @@
-﻿namespace Axent.Core;
+﻿namespace Axent.Abstractions;
 
 /// <summary>
 /// Use this type when your IRequestHandler should return void